chore: changed all component and utility files to to ts equivalent file tyoes
