+++ src/utilities/checkClearLocalStorage.ts
