+++ src/utilities/deleteAllToDos.ts
