--- src/utilities/deleteToDo.js
@@ -1,9 +1,0 @@
-export default function deleteToDo(allToDos, setAllToDos, itemKey) {
-    const copyAllToDos = {...allToDos};
-    delete copyAllToDos[itemKey];
-    setAllToDos(copyAllToDos);
-
-    // deleting to-do from localStorage
-
-    localStorage.removeItem(itemKey);
-  }
